decouple session class from existing Connection
ensure maintenance of current APIs of Connection while delegating
responsibility
